Refactor frontend code
