+++ src/lib/components/layout/index.ts
@@ -1,3 +1,4 @@
 export { default as Sidebar } from './Sidebar.svelte'
 export { default as Toolbar } from './Toolbar.svelte'
 export { default as RightSidebar } from './RightSidebar.svelte'
+export { default as OrchestratorLayer } from './OrchestratorLayer.svelte'
