+++ src/lib/hooks/useAppSetup.ts
@@ -1,0 +1,603 @@
+import { tick } from 'svelte'
+import { get } from 'svelte/store'
+import type { ActiveView, DuplicateTrack, Playlist, SettingsPage, Track, UsbDevice } from '$lib/types'
+import {
+	appStore,
+	libraryStore,
+	sortedTracks,
+	displayedTracks,
+	playerStore,
+	currentTrack,
+	tagsStore,
+	playlistsStore,
+	uiStore,
+	activeView,
+	selectedTrackIds,
+	selectedReleaseIds,
+	settingsStore,
+	continuousPlayback,
+	devicesStore,
+	missingTracksStore,
+	missingTrackIds,
+	sortedReleases,
+	expandedReleaseIds,
+	discoveryStore,
+	updaterStore,
+	previewInfo,
+} from '$lib/stores'
+import { tagFilterMode } from '$lib/stores/ui'
+import { recentlyToggledMixedTags } from '$lib/stores/ui'
+import { syncStore } from '$lib/stores/sync'
+import { toastStore } from '$lib/stores/toast'
+import { exportStore } from '$lib/stores/export'
+import { dismissSplash } from '$lib/stores/splash'
+import { discoveryPlaylistStore } from '$lib/stores/discoveryPlaylist'
+import {
+	createTagController,
+	createTrackController,
+	createDeviceController,
+	createExportController,
+	createPlaylistController,
+} from '$lib/controllers'
+import { useAppInitialization } from './useAppInitialization'
+import { useKeyboardShortcuts } from './useKeyboardShortcuts'
+import { useMenuActions } from './useMenuActions'
+import { useMediaKeys } from './useMediaKeys'
+import { useDragDropCoordination } from './useDragDropCoordination'
+import { translate } from '$lib/i18n'
+import * as playlistsApi from '$lib/api/playlists'
+
+// =============================================================================
+// Types
+// =============================================================================
+
+export interface AppSetupConfig {
+	getPlaylists: () => Playlist[]
+	getDevices: () => UsbDevice[]
+	getSelectedPlaylistId: () => string | null
+	getSelectedFolderId: () => string | null
+	getSelectedTagIds: () => string[]
+	getModalOrchestrator: () => ModalOrchestratorRef | undefined
+	handleViewChange: (view: ActiveView) => void
+	setShowAddReleaseModal: () => void
+	setIsDragOver: (dragOver: boolean) => void
+}
+
+interface ModalOrchestratorRef {
+	isModalOpen: () => boolean
+	openSettingsModal: (tab?: SettingsPage) => void
+	openCreatePlaylistModal: (parentId: string | null) => void
+	openCreateFolderModal: (parentId: string | null) => void
+	openCreateSmartPlaylistModal: (parentId: string | null, context?: ActiveView) => void
+	openEditSmartPlaylistModal: (playlist: Playlist) => void
+	openRenamePlaylistModal: (playlist: Playlist) => void
+	openDeletePlaylistModal: (playlist: Playlist, hasChildren: boolean) => void
+	openDeletePlaylistBulkModal: (playlists: Playlist[]) => void
+	openMoveConflictModal: (playlist: Playlist, conflict: Playlist, targetId: string | null) => void
+	openRelocateModal: (track: Track) => void
+	openRemoveFromPlaylistModal: (trackIds: string[], playlistId: string) => void
+	openRemoveFromLibraryModal: (trackIds: string[]) => void
+	openRemoveDiscoveryReleasesModal: (releaseIds: string[]) => void
+	openRemoveDiscoveryReleasesFromPlaylistModal: (releaseIds: string[], playlistId: string) => void
+	openDuplicateTrackModal: (
+		duplicates: DuplicateTrack[],
+		onComplete: (updatedTracks: Track[], newTracks: Track[], replacedTrackIds: string[]) => void
+	) => void
+	openDeviceInfoModal: (device: UsbDevice) => void
+	openReformatDeviceModal: (device: UsbDevice) => void
+	openExportToDeviceModal: (device: UsbDevice) => void
+	openExportPlaylistModal: (playlist: Playlist) => void
+	openQuickExportModal: () => void
+	openExportFailureModal: (error: string, deviceId: string, mountPoint: string, filesCopied: number) => void
+}
+
+export interface AppSetupResult {
+	tagController: ReturnType<typeof createTagController>
+	trackController: ReturnType<typeof createTrackController>
+	deviceController: ReturnType<typeof createDeviceController>
+	exportController: ReturnType<typeof createExportController>
+	playlistController: ReturnType<typeof createPlaylistController>
+	playNextTrack: () => void
+	playPreviousTrack: () => void
+	onMountSetup: () => Promise<() => void>
+	setupDragDrop: () => () => void
+}
+
+// =============================================================================
+// Hook
+// =============================================================================
+
+export function createAppSetup(config: AppSetupConfig): AppSetupResult {
+	const {
+		getPlaylists,
+		getDevices,
+		getSelectedPlaylistId,
+		getSelectedFolderId,
+		getSelectedTagIds,
+		getModalOrchestrator,
+		handleViewChange,
+		setShowAddReleaseModal,
+		setIsDragOver,
+	} = config
+
+	// =========================================================================
+	// Controllers
+	// =========================================================================
+
+	const tagController = createTagController({
+		tagsStore,
+		libraryStore,
+		discoveryStore,
+		uiStore,
+		getSelectedTagIds,
+		getSelectedPlaylistId,
+		getTagFilterMode: () => get(tagFilterMode),
+		getSelectedTrackIds: () => get(selectedTrackIds),
+		getSelectedReleaseIds: () => get(selectedReleaseIds),
+		getRecentlyToggledMixedTags: () => get(recentlyToggledMixedTags),
+		getActiveView: () => get(activeView),
+	})
+
+	const trackController = createTrackController(
+		{
+			playerStore,
+			libraryStore,
+			playlistsStore,
+			missingTracksStore,
+			uiStore,
+			toastStore,
+			getSelectedPlaylistId,
+			getPlaylists,
+			getMissingTrackIds: () => get(missingTrackIds),
+		},
+		{
+			openRelocateModal: (track) => getModalOrchestrator()?.openRelocateModal(track),
+			openRemoveFromPlaylistModal: (trackIds, playlistId) =>
+				getModalOrchestrator()?.openRemoveFromPlaylistModal(trackIds, playlistId),
+			openRemoveFromLibraryModal: (trackIds) => getModalOrchestrator()?.openRemoveFromLibraryModal(trackIds),
+			openDuplicateTrackModal: (duplicates, onComplete) =>
+				getModalOrchestrator()?.openDuplicateTrackModal(duplicates, onComplete),
+		}
+	)
+
+	const deviceController = createDeviceController(
+		{ devicesStore, settingsStore, toastStore },
+		{
+			openDeviceInfoModal: (device) => getModalOrchestrator()?.openDeviceInfoModal(device),
+			openReformatDeviceModal: (device) => getModalOrchestrator()?.openReformatDeviceModal(device),
+		}
+	)
+
+	const exportController = createExportController(
+		{
+			exportStore,
+			toastStore,
+			getDevices,
+			getPlaylists,
+		},
+		{
+			openExportToDeviceModal: (device) => getModalOrchestrator()?.openExportToDeviceModal(device),
+			openExportPlaylistModal: (playlist) => getModalOrchestrator()?.openExportPlaylistModal(playlist),
+			openQuickExportModal: () => getModalOrchestrator()?.openQuickExportModal(),
+			openExportFailureModal: (error, deviceId, mountPoint, filesCopied) =>
+				getModalOrchestrator()?.openExportFailureModal(error, deviceId, mountPoint, filesCopied),
+		}
+	)
+
+	const playlistController = createPlaylistController(
+		{
+			playlistsStore,
+			libraryStore,
+			uiStore,
+			toastStore,
+			getPlaylists,
+			getSelectedPlaylistId,
+			getSelectedFolderId,
+			getSelectedTagIds,
+			getTagFilterMode: () => get(tagFilterMode),
+			onDiscoveryPlaylistSelected: async (playlistId) => {
+				const playlist = getPlaylists().find((p) => p.id === playlistId)
+				const releases = playlist?.is_smart
+					? await playlistsApi.getSmartPlaylistReleases(playlistId)
+					: await playlistsStore.getPlaylistReleases(playlistId)
+				discoveryPlaylistStore.cacheAndSet(playlistId, releases)
+			},
+		},
+		{
+			openCreatePlaylistModal: (parentId) => getModalOrchestrator()?.openCreatePlaylistModal(parentId),
+			openCreateFolderModal: (parentId) => getModalOrchestrator()?.openCreateFolderModal(parentId),
+			openCreateSmartPlaylistModal: (parentId, context) =>
+				getModalOrchestrator()?.openCreateSmartPlaylistModal(parentId, context),
+			openEditSmartPlaylistModal: (playlist) => getModalOrchestrator()?.openEditSmartPlaylistModal(playlist),
+			openRenamePlaylistModal: (playlist) => getModalOrchestrator()?.openRenamePlaylistModal(playlist),
+			openDeletePlaylistModal: (playlist, hasChildren) =>
+				getModalOrchestrator()?.openDeletePlaylistModal(playlist, hasChildren),
+			openMoveConflictModal: (playlist, conflict, targetId) =>
+				getModalOrchestrator()?.openMoveConflictModal(playlist, conflict, targetId),
+		}
+	)
+
+	// =========================================================================
+	// Track Navigation
+	// =========================================================================
+
+	function playNextTrack() {
+		const preview = get(previewInfo)
+		if (preview) {
+			const nextIndex = preview.trackIndex + 1
+			if (nextIndex < preview.release.tracks.length) {
+				playerStore.playPreview(preview.release, nextIndex)
+			} else {
+				playerStore.stop()
+			}
+			return
+		}
+		const id = get(currentTrack)?.id
+		if (!id) return
+		const tracks = get(displayedTracks)
+		const idx = tracks.findIndex((t) => t.id === id)
+		if (idx >= 0 && idx < tracks.length - 1) trackController.play(tracks[idx + 1])
+	}
+
+	function playPreviousTrack() {
+		const preview = get(previewInfo)
+		if (preview) {
+			const prevIndex = preview.trackIndex - 1
+			if (prevIndex >= 0) {
+				playerStore.playPreview(preview.release, prevIndex)
+			}
+			return
+		}
+		const id = get(currentTrack)?.id
+		if (!id) return
+		const tracks = get(displayedTracks)
+		const idx = tracks.findIndex((t) => t.id === id)
+		if (idx > 0) trackController.play(tracks[idx - 1])
+	}
+
+	// =========================================================================
+	// Shared Handlers (used by both keyboard shortcuts and menu actions)
+	// =========================================================================
+
+	const handlers = {
+		playPause: () => playerStore.togglePlayPause(),
+		stop: () => playerStore.stop(),
+		seekForward: () => playerStore.seekRelative(10000),
+		seekBackward: () => playerStore.seekRelative(-10000),
+		fineSeekForward: () => playerStore.seekRelative(1000),
+		fineSeekBackward: () => playerStore.seekRelative(-1000),
+		volumeUp: () => playerStore.adjustVolume(0.1),
+		volumeDown: () => playerStore.adjustVolume(-0.1),
+		toggleMute: () => playerStore.toggleMute(),
+
+		selectAll: () => {
+			if (get(activeView) === 'discovery') {
+				uiStore.setSelectedReleases(new Set(get(sortedReleases).map((r) => r.id)))
+			} else {
+				uiStore.setSelectedTracks(new Set(get(sortedTracks).map((t) => t.id)))
+			}
+		},
+
+		openSettings: (tab?: SettingsPage) => getModalOrchestrator()?.openSettingsModal(tab),
+
+		quickExport: () => {
+			if (getDevices().length > 0) getModalOrchestrator()?.openQuickExportModal()
+		},
+
+		jumpToPlayingTrack: () => {
+			const track = get(currentTrack)
+			if (!track) return
+			if (getSelectedPlaylistId()) playlistController.handleLibraryClick()
+			uiStore.selectTrack(track.id)
+		},
+
+		toggleView: () => {
+			if (getModalOrchestrator()?.isModalOpen()) return
+			const next = get(activeView) === 'library' ? 'discovery' : 'library'
+			handleViewChange(next)
+		},
+
+		import: async () => {
+			if (get(activeView) !== 'library') {
+				handleViewChange('library')
+				await tick()
+			}
+			trackController.handleImport()
+		},
+
+		addRelease: async () => {
+			if (get(activeView) !== 'discovery') {
+				handleViewChange('discovery')
+				await tick()
+			}
+			setShowAddReleaseModal()
+		},
+
+		refreshMetadata: async () => {
+			if (get(activeView) !== 'discovery') return
+			const ids = [...get(selectedReleaseIds)]
+			if (ids.length === 0) return
+			await Promise.all(ids.map((id) => discoveryStore.refreshMetadata(id)))
+		},
+	}
+
+	// =========================================================================
+	// Mount Setup
+	// =========================================================================
+
+	async function onMountSetup(): Promise<() => void> {
+		const splashStartTime = Date.now()
+		const minDisplayTime = 700
+
+		await exportStore.startListening()
+
+		const cleanupApp = await useAppInitialization({
+			stores: { appStore, libraryStore, tagsStore, playlistsStore, settingsStore, devicesStore, syncStore },
+			toastStore,
+			onExternalFileDrop: trackController.handleExternalFileDrop,
+			onDragStateChange: (dragOver) => setIsDragOver(dragOver),
+		})
+
+		const cleanupKeyboard = useKeyboardShortcuts({
+			isModalOpen: () => getModalOrchestrator()?.isModalOpen() ?? false,
+			onPlayPause: handlers.playPause,
+			onFocusSearch: () => {
+				const searchInput = document.querySelector('input[type="search"]') as HTMLInputElement
+				searchInput?.focus()
+			},
+			onClearSelection: () => uiStore.clearSelection(),
+			onSelectAll: handlers.selectAll,
+			onOpenSettings: handlers.openSettings,
+			onNewPlaylist: () => playlistController.handleCreatePlaylist(),
+			onNewFolder: () => playlistController.handleCreateFolder(),
+			onImport: handlers.import,
+			onDeleteSelected: () => {
+				const treeIds = get(uiStore).selectedTreeIds
+				if (treeIds.size > 1) {
+					const selected = getPlaylists().filter((p) => treeIds.has(p.id))
+					if (selected.length > 0) {
+						getModalOrchestrator()?.openDeletePlaylistBulkModal(selected)
+					}
+					return true
+				}
+
+				const playlistId = getSelectedPlaylistId()
+				const playlists = getPlaylists()
+				const currentPlaylist = playlistId ? playlists.find((p) => p.id === playlistId) : null
+				if (currentPlaylist?.is_smart) return false
+
+				if (get(activeView) === 'discovery') {
+					const releaseIds = get(selectedReleaseIds)
+					if (releaseIds.size > 0) {
+						if (playlistId) {
+							getModalOrchestrator()?.openRemoveDiscoveryReleasesFromPlaylistModal(Array.from(releaseIds), playlistId)
+						} else {
+							getModalOrchestrator()?.openRemoveDiscoveryReleasesModal(Array.from(releaseIds))
+						}
+						return true
+					}
+				}
+				const ids = [...get(selectedTrackIds)]
+				if (ids.length > 0) {
+					if (playlistId) {
+						getModalOrchestrator()?.openRemoveFromPlaylistModal(ids, playlistId)
+					} else {
+						getModalOrchestrator()?.openRemoveFromLibraryModal(ids)
+					}
+				} else if (playlistId) {
+					if (currentPlaylist) playlistController.handlePlaylistDelete(currentPlaylist)
+				} else {
+					const folderId = getSelectedFolderId()
+					if (folderId) {
+						const folder = playlists.find((p) => p.id === folderId)
+						if (folder) playlistController.handlePlaylistDelete(folder)
+					}
+				}
+				return true
+			},
+			onPlaySelected: () => {
+				if (get(activeView) === 'discovery') {
+					const releaseIds = get(selectedReleaseIds)
+					if (releaseIds.size > 0) {
+						const releases = get(sortedReleases)
+						expandedReleaseIds.toggleSelection(
+							[...releaseIds],
+							(id) => (releases.find((r) => r.id === id)?.tracks.length ?? 0) > 0
+						)
+					}
+					return
+				}
+				const selectedIds = get(selectedTrackIds)
+				if (selectedIds.size > 0) {
+					const firstSelectedId = [...selectedIds][0]
+					const track = get(displayedTracks).find((t) => t.id === firstSelectedId)
+					if (track) trackController.play(track)
+				}
+			},
+			onSeekBackward: handlers.seekBackward,
+			onSeekForward: handlers.seekForward,
+			onFineSeekBackward: handlers.fineSeekBackward,
+			onFineSeekForward: handlers.fineSeekForward,
+			onPreviousTrack: playPreviousTrack,
+			onNextTrack: playNextTrack,
+			onVolumeUp: handlers.volumeUp,
+			onVolumeDown: handlers.volumeDown,
+			onToggleMute: handlers.toggleMute,
+			onSelectPreviousTrack: () => {
+				if (get(activeView) === 'discovery') {
+					const releases = get(sortedReleases)
+					if (releases.length === 0) return
+					const ids = get(selectedReleaseIds)
+					if (ids.size === 0) {
+						uiStore.selectRelease(releases[releases.length - 1].id)
+					} else {
+						const firstId = [...ids][0]
+						const idx = releases.findIndex((r) => r.id === firstId)
+						if (idx > 0) uiStore.selectRelease(releases[idx - 1].id)
+					}
+					return
+				}
+				const tracks = get(displayedTracks)
+				if (tracks.length === 0) return
+				const ids = get(selectedTrackIds)
+				if (ids.size === 0) {
+					uiStore.selectTrack(tracks[tracks.length - 1].id)
+				} else {
+					const firstId = [...ids][0]
+					const idx = tracks.findIndex((t) => t.id === firstId)
+					if (idx > 0) uiStore.selectTrack(tracks[idx - 1].id)
+				}
+			},
+			onSelectNextTrack: () => {
+				if (get(activeView) === 'discovery') {
+					const releases = get(sortedReleases)
+					if (releases.length === 0) return
+					const ids = get(selectedReleaseIds)
+					if (ids.size === 0) {
+						uiStore.selectRelease(releases[0].id)
+					} else {
+						const lastId = [...ids].pop()
+						const idx = releases.findIndex((r) => r.id === lastId)
+						if (idx >= 0 && idx < releases.length - 1) uiStore.selectRelease(releases[idx + 1].id)
+					}
+					return
+				}
+				const tracks = get(displayedTracks)
+				if (tracks.length === 0) return
+				const ids = get(selectedTrackIds)
+				if (ids.size === 0) {
+					uiStore.selectTrack(tracks[0].id)
+				} else {
+					const lastId = [...ids].pop()
+					const idx = tracks.findIndex((t) => t.id === lastId)
+					if (idx >= 0 && idx < tracks.length - 1) uiStore.selectTrack(tracks[idx + 1].id)
+				}
+			},
+			onQuickExport: handlers.quickExport,
+			onJumpToPlayingTrack: handlers.jumpToPlayingTrack,
+			onToggleView: handlers.toggleView,
+			onAddRelease: handlers.addRelease,
+			onRefreshMetadata: handlers.refreshMetadata,
+		})
+
+		const cleanupMenu = await useMenuActions({
+			onImport: handlers.import,
+			onAddRelease: handlers.addRelease,
+			onCreatePlaylist: playlistController.handleCreatePlaylist,
+			onCreateFolder: playlistController.handleCreateFolder,
+			onSelectAll: handlers.selectAll,
+			onPlayPause: handlers.playPause,
+			onStop: handlers.stop,
+			onNextTrack: playNextTrack,
+			onPreviousTrack: playPreviousTrack,
+			onSeekForward: handlers.seekForward,
+			onSeekBackward: handlers.seekBackward,
+			onFineSeekForward: handlers.fineSeekForward,
+			onFineSeekBackward: handlers.fineSeekBackward,
+			onVolumeUp: handlers.volumeUp,
+			onVolumeDown: handlers.volumeDown,
+			onToggleMute: handlers.toggleMute,
+			onOpenSettings: handlers.openSettings,
+			onQuickExport: handlers.quickExport,
+			onJumpToPlayingTrack: handlers.jumpToPlayingTrack,
+			onToggleView: handlers.toggleView,
+			onToggleEditor: () => uiStore.toggleRightSidebar(),
+			onExpandAllReleases: () => {
+				const releases = get(sortedReleases)
+				expandedReleaseIds.expandAll(releases.filter((r) => r.tracks.length > 0).map((r) => r.id))
+			},
+			onCollapseAllReleases: () => expandedReleaseIds.collapseAll(),
+			onRefreshMetadata: handlers.refreshMetadata,
+		})
+
+		const cleanupMediaKeys = await useMediaKeys({
+			onPlayPause: handlers.playPause,
+			onNextTrack: playNextTrack,
+			onPreviousTrack: playPreviousTrack,
+		})
+
+		playerStore.onTrackEnd(() => {
+			if (get(continuousPlayback)) {
+				playNextTrack()
+			}
+		})
+
+		const elapsed = Date.now() - splashStartTime
+		if (elapsed < minDisplayTime) {
+			await new Promise((r) => setTimeout(r, minDisplayTime - elapsed))
+		}
+
+		if (get(activeView) === 'discovery') {
+			discoveryStore.loadReleases()
+		}
+
+		updaterStore.check(true)
+		const updateInterval = setInterval(() => updaterStore.check(true), 60 * 60 * 1000)
+
+		dismissSplash()
+
+		return () => {
+			cleanupApp()
+			cleanupKeyboard()
+			cleanupMenu()
+			cleanupMediaKeys()
+			playerStore.onTrackEnd(null)
+			exportStore.stopListening()
+			clearInterval(updateInterval)
+		}
+	}
+
+	// =========================================================================
+	// Drag-Drop Setup
+	// =========================================================================
+
+	function setupDragDrop(): () => void {
+		return useDragDropCoordination({
+			getPlaylists,
+			getDevices,
+			onTracksDropOnPlaylist: trackController.handleTracksDropOnPlaylist,
+			onReleasesDropOnPlaylist: async (playlistId: string, releaseIds: string[]) => {
+				await playlistsStore.addReleases(playlistId, releaseIds)
+			},
+			onPlaylistMove: playlistController.handlePlaylistDragMove,
+			onPlaylistExportToDevice: exportController.handlePlaylistDropOnDevice,
+			onTagDropOnTrack: async (tagId: string, trackId: string) => {
+				const trackIds = get(selectedTrackIds).has(trackId) ? Array.from(get(selectedTrackIds)) : [trackId]
+				await tagsStore.assignTags(trackIds, [tagId])
+				const playlistId = getSelectedPlaylistId()
+				if (playlistId) {
+					await libraryStore.loadPlaylistTracks(playlistId)
+				} else {
+					await libraryStore.loadTracks()
+				}
+			},
+			onTagDropOnRelease: async (tagId: string, releaseId: string) => {
+				const releaseIds = get(selectedReleaseIds).has(releaseId) ? Array.from(get(selectedReleaseIds)) : [releaseId]
+				await discoveryStore.assignTags(releaseIds, [tagId])
+			},
+			onTagDropOnCategory: async (tagId: string, _sourceCategoryId: string, targetCategoryId: string) => {
+				try {
+					await tagsStore.moveTag(tagId, targetCategoryId)
+					libraryStore.updateTagCategory(tagId, targetCategoryId)
+					discoveryStore.updateTagCategory(tagId, targetCategoryId)
+					discoveryPlaylistStore.updateTagCategory(tagId, targetCategoryId)
+				} catch (error) {
+					const message = error instanceof Error ? error.message : get(translate)('errors.tagNameConflict')
+					toastStore.error(message)
+				}
+			},
+		})
+	}
+
+	return {
+		tagController,
+		trackController,
+		deviceController,
+		exportController,
+		playlistController,
+		playNextTrack,
+		playPreviousTrack,
+		onMountSetup,
+		setupDragDrop,
+	}
+}
